fix: restliche Compile-Fehler beheben – APK baut erfolgreich (61 MB)
Lokal mit Flutter 3.35.5 verifiziert: flutter analyze 0 Fehler,
flutter build apk --release erfolgreich (app-release.apk, 61,0 MB).

- soundtrack_provider.dart: doppelte switch-Labels (erwachsen/reife)
  entfernt und fehlenden GamePhase.formung-Zweig ergänzt (beide Switches)
- hindernis.dart: kaputtes Latin-1-Byte (ü) nach UTF-8 korrigiert
- entstehungs_controller.dart: GamePhase-Import ergänzt
- entstehungs_spiel.dart: flame/components-Import für Anchor ergänzt
- erweiterungen.dart: karmaBegrenzt() gibt double statt num zurück
- reinkarnations_screen.dart: Icons.check_small -> Icons.check
- genesis_app.dart: CardTheme -> CardThemeData (Flutter 3.35 API)

+++ lib/core/utils/erweiterungen.dart
@@ -48,7 +48,7 @@ extension DoubleSpielErweiterungen on double {
       AppFarben.karmaFarbeInterpoliert(karmaZuProzent());
 
   /// Begrenzt diesen Wert auf den gültigen Karma-Bereich [-100, +100].
-  double karmaBegrenzt() => clamp(kMinKarma, kMaxKarma);
+  double karmaBegrenzt() => clamp(kMinKarma, kMaxKarma).toDouble();
 
   // ── Allgemeine Hilfsmethoden ─────────────────────────────────────────────
